backend/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
sqlmodel>=0.0.8
sentence-transformers>=2.2.2
torch>=2.0.0
faiss-cpu>=1.7.4.post2
pandas>=1.5
numpy>=1.24
requests>=2.31
fhir.resources>=6.0.0
openpyxl>=3.1.0
python-multipart>=0.0.6
scikit-learn>=1.2.2
